commit after Client Controller

+++ Repositories/AdvertisementRepository.cs
@@ -6,17 +6,21 @@ using NewspaperAdvertisementManagementSystem.Contexts;
 using Microsoft.EntityFrameworkCore;
 using System.Linq;
 using System;
+using Microsoft.AspNetCore.Http;
+using System.IO;
+using Microsoft.AspNetCore.Hosting;
 
 namespace NewspaperAdvertisementManagementSystem.Repositories
 {
     public class AdvertisementRepository : IAdvertisementRepository
     {
         private readonly ApplicationDbContext _context;
+        private readonly IWebHostEnvironment _hostEnvironment;
 
-        public AdvertisementRepository(ApplicationDbContext context)
+        public AdvertisementRepository(ApplicationDbContext context, IWebHostEnvironment hostEnvironment)
         {
             this._context = context;
-
+            this._hostEnvironment = hostEnvironment;
         }
         public async Task<List<Advertisement>> GetAdvertisements()
         {
@@ -30,12 +34,41 @@ namespace NewspaperAdvertisementManagementSystem.Repositories
 
         public async Task<int> AddAdvertisement(Advertisement advertisement)
         {
+            advertisement.AdvertisementImageName = await SaveImage(advertisement.AdvertisementImageFile);
             var result = await _context.Advertisements.AddAsync(advertisement);
 
             await _context.SaveChangesAsync();
 
             return result.Entity.AdvertisementId;
         }
+
+
+        public async Task<string> SaveImage(IFormFile imageFile)
+        {
+            string imageName = new String(Path.GetFileNameWithoutExtension(imageFile.FileName).Take(10).ToArray()).Replace(' ', '-');
+            imageName = imageName + DateTime.Now.ToString("yymmssfff") + Path.GetExtension(imageFile.FileName);
+
+            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, "Images//Advertisements", imageName);
+
+            using (var fileStream = new FileStream(imagePath, FileMode.Create))
+            {
+                await imageFile.CopyToAsync(fileStream);
+            }
+
+            return imageName;
+
+
+
+        }
+
+
+
+
+
+
+
+
+
 
         public async Task<Advertisement> UpdateAdvertisementStatus(int AdvertisementId)
         {
@@ -69,13 +102,20 @@ namespace NewspaperAdvertisementManagementSystem.Repositories
 
         }
 
-        public async Task DeleteAdvertisement(int AdvertisementId)
+        public async Task<string> DeleteAdvertisement(int AdvertisementId)
         {
             var result = await _context.Advertisements.FirstOrDefaultAsync(x => x.AdvertisementId == AdvertisementId);
 
-            _context.Advertisements.Remove(result);
+            if (result != null)
+            {
+                _context.Advertisements.Remove(result);
 
-            await _context.SaveChangesAsync();
+                await _context.SaveChangesAsync();
+
+                return "SuccessfullyDeleted";
+
+            }
+            return null;
 
         }
